Revert the addition of get_input_report
The underlying hid_get_input_report API has not yet reached a HIDAPI
release, and attempting to wrap it causes an "undefined symbol" when the
hid module is imported.

Revert the upstream changes that add this new method.

Note: as hidraw.pyx is a symlink in the repository, but a regular file
in the sdist that is used in this package, it is not sufficient to use
the patch generated by git revert.

Related: FS#67103 <https://bugs.archlinux.org/task/67103>