go: stop using --no-clean with make.bash
The script now says:

	warning: --no-clean is deprecated and has no effect; use 'go install std cmd' instead

That is, if we are building from source for the first time,
we want "./make.bash", which will do that.
If we already have Go built but some packages need rebuilding,
then "go install std cmd" will do the job.

We're in the first of those two scenarios.